requirements.txt
Django==4.2.9
django-treebeard==4.7
